Move some mixin class to appropriate folder
