+++ common/src/main/resources/vs_ship_assembler-common.mixins.json
@@ -3,8 +3,8 @@
   "package": "io.github.priestoffern.vs_ship_assembler.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "GameRendererMixin",
-    "ShipAssemblyMixin"
+    "features.GameRendererMixin",
+    "features.ShipAssemblyMixin"
   ],
   "injectors": {
     "defaultRequire": 1
